+++ Jim/Forms/ProductCategorySetupForm.cs
@@ -54,15 +54,19 @@ namespace Jim.Forms
 
         private void barButtonDelete_ItemClick(object sender, DevExpress.XtraBars.ItemClickEventArgs e)
         {
-            var row = gridView.GetFocusedRow() as ProductCategoryModel;
-            if (row.ProductCategoryID != null && row.ProductCategoryID != Guid.Empty)
+            DialogResult res = XtraMessageBox.Show(String.Format("Σίγουρα θέλετε να διαγραφεί;"), "", MessageBoxButtons.YesNo, MessageBoxIcon.Warning);
+            if (res == DialogResult.Yes)
             {
-                using (var repository = new ProductCategoryRepository())
+                var row = gridView.GetFocusedRow() as ProductCategoryModel;
+                if (row.ProductCategoryID != null && row.ProductCategoryID != Guid.Empty)
                 {
-                    repository.Delete(row.ProductCategoryID);
+                    using (var repository = new ProductCategoryRepository())
+                    {
+                        repository.Delete(row.ProductCategoryID);
+                    }
                 }
+                this.gridControl.EmbeddedNavigator.Buttons.DoClick(this.gridControl.EmbeddedNavigator.Buttons.Remove);
             }
-            this.gridControl.EmbeddedNavigator.Buttons.DoClick(this.gridControl.EmbeddedNavigator.Buttons.Remove);
         }
 
         private void gridView_CellValueChanged(object sender, DevExpress.XtraGrid.Views.Base.CellValueChangedEventArgs e)
