+++ Jim/Forms/ProductSetupForm.cs
@@ -60,15 +60,20 @@ namespace Jim.Forms
 
         private void barButtonDelete_ItemClick(object sender, DevExpress.XtraBars.ItemClickEventArgs e)
         {
-            var row = gridView.GetFocusedRow() as ProductModel;
-            if (row.ProductID != null && row.ProductID != Guid.Empty)
+
+            DialogResult res = XtraMessageBox.Show(String.Format("Σίγουρα θέλετε να διαγραφεί;"), "", MessageBoxButtons.YesNo, MessageBoxIcon.Warning);
+            if (res == DialogResult.Yes)
             {
-                using (var repository = new ProductsRepository())
+                var row = gridView.GetFocusedRow() as ProductModel;
+                if (row.ProductID != null && row.ProductID != Guid.Empty)
                 {
-                    repository.Delete(row.ProductID);
+                    using (var repository = new ProductsRepository())
+                    {
+                        repository.Delete(row.ProductID);
+                    }
                 }
+                this.gridControl.EmbeddedNavigator.Buttons.DoClick(this.gridControl.EmbeddedNavigator.Buttons.Remove);
             }
-            this.gridControl.EmbeddedNavigator.Buttons.DoClick(this.gridControl.EmbeddedNavigator.Buttons.Remove);
         }
         private void gridView_CellValueChanged(object sender, DevExpress.XtraGrid.Views.Base.CellValueChangedEventArgs e)
         {
